requirements/base.txt
fastapi[standard],
SQLAlchemy
psycopg2
bcrypt
pydantic-settings
python-dotenv
passlib[bcrypt]
openssl
PyJWT